+++ src/main/java/com/kodilla/ecommercee/controller/CartController.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.controller;
+
+public class CartController {
+}
