+++ src/main/java/com/kodilla/ecommercee/domain/User.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.domain;
+
+public class User {
+}
